fix breakout actions dtype: double* → float* (vecenv sync broke it)
upstream 4.0 sync (ad22a6e5e) changed vecenv actions from double to float but breakout.h wasn't updated. env read float data through double pointer — garbage actions, no learning. bisect confirmed this commit as the regression: score 734→3.1.

+++ pufferlib/ocean/breakout/breakout.h
@@ -40,7 +40,7 @@ typedef struct Breakout {
     Client* client;
     Log log;
     float* observations;
-    double* actions;
+    float* actions;
     float* rewards;
     float* terminals;
     int num_agents;
@@ -121,7 +121,7 @@ void init(Breakout* env) {
 void allocate(Breakout* env) {
     init(env);
     env->observations = (float*)calloc(11 + env->num_bricks, sizeof(float));
-    env->actions = (double*)calloc(1, sizeof(double));
+    env->actions = (float*)calloc(1, sizeof(float));
     env->rewards = (float*)calloc(1, sizeof(float));
     env->terminals = (float*)calloc(1, sizeof(float));
 }
